redid query ahk2

+++ install.ps1
@@ -4,15 +4,19 @@ Set-Location "$env:APPDATA\Microsoft\Windows\Start Menu\Programs\Startup"
 
 # Remove the old AHK file
 # Prompt the user for confirmation before deleting .ahk and .ahk2 files
-$confirmation = Read-Host "Do you want to delete all .ahk and .ahk2 files in the directory? (Y/N)"
-if ($confirmation -eq 'Y') {
-    Remove-Item *.ahk, *.ahk2 -Force
-    Write-Host "Files deleted."
-} else {
-    Write-Host "Deletion cancelled."
-    
-    Write-Host "Exiting script."
-    exit
+# Check if the folder contains .ahk or .ahk2 files
+$ahkFiles = Get-ChildItem -Filter *.ahk
+$ahk2Files = Get-ChildItem -Filter *.ahk2
+if ($ahkFiles -or $ahk2Files) {
+    $confirmation = Read-Host "The folder contains .ahk or .ahk2 files. Do you want to delete them? (Y/N)"
+    if ($confirmation -eq 'Y') {
+        Remove-Item *.ahk, *.ahk2 -Force
+        Write-Host "Files deleted."
+    } else {
+        Write-Host "Deletion cancelled."
+        Write-Host "Exiting script."
+        exit
+    }
 }
 
 # Download all the seperate AHK files
